requirements.txt
bottle
gunicorn
python-dotenv==0.15.0
sentry-sdk==0.16.2